requirements.txt
airtable-python-wrapper==0.8.0
astroid==2.2.5
certifi==2018.1.18
chardet==3.0.4
colorama==0.4.1
Django==2.0
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-filter==2.1.0
django-filters==0.2.1
django-phone-field==1.7.1
django-static-precompiler==1.8.2
idna==2.6
isort==4.3.20
lazy-object-proxy==1.4.1
libsass==0.19.1
mccabe==0.6.1
Pillow==6.0.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pylint-protobuf==0.3
pytz==2017.3
requests==2.18.4
rjsmin==1.1.0
six==1.11.0
typed-ast==1.3.5
urllib3==1.22
virtualenv==16.6.0
wrapt==1.11.1
